requirements.txt
amqp==5.1.1
aniso8601==7.0.0
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-rest-swagger==2.2.0
django-utils-six==2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
fastdiff==0.3.0
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
loguru==0.7.0
Markdown==3.4.4
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
Pillow==10.0.0
promise==2.3
prompt-toolkit==3.0.39
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
Rx==1.6.3
simplejson==3.19.1
singledispatch==4.0.0
six==1.16.0
snapshottest==0.6.0
sqlparse==0.4.4
termcolor==2.3.0
text-unidecode==1.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wasmer==1.1.0
wasmer-compiler-cranelift==1.1.0
wcwidth==0.2.6
win32-setctime==1.1.0
